remove bad previous approach and just do login flow with states
