--- website/src/routes/log-in/passkey/+page.svelte
@@ -1,5 +1,0 @@
-<script>
-    import PasskeyLogin from '$components/login/passkey-login.svelte';
-</script>
-
-<PasskeyLogin />
