--- website/src/routes/log-in/password/+page.svelte
@@ -1,5 +1,0 @@
-<script>
-    import PasswordLogin from '$components/login/password-login.svelte';
-</script>
-
-<PasswordLogin />
